all entities now have index show and delete methods. also changed many to one anotations

+++ src/main/java/repositories/PurchaseRequestLineItemRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import entities.PurchaseRequestLineItem;
 
-public interface PurchaseRequestLineItemRepository extends CrudRepository<PurchaseRequestLineItem, Integer> {
+public interface PurchaseRequestLineItemRepository extends CrudRepository<PurchaseRequestLineItem, Long> {
 
 }
