+++ src/main/java/repositories/PurchaseRequestRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import entities.PurchaseRequest;
 
-public interface PurchaseRequestRepository extends CrudRepository<PurchaseRequest, Integer> {
+public interface PurchaseRequestRepository extends CrudRepository<PurchaseRequest, Long> {
 
 }
