+++ src/main/java/repositories/VendorRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import entities.Vendor;
 
-public interface VendorRepository extends CrudRepository<Vendor, Integer> {
+public interface VendorRepository extends CrudRepository<Vendor, Long> {
 
 }
